put into components and pages
