+++ web/src/components/DraggableCard.tsx
@@ -1,0 +1,1 @@
+
