+++ web/src/components/index.ts
@@ -1,0 +1,8 @@
+// NB: Adding this index allows a single line import in App.tsx making view easier
+
+export { default as Header } from "./Header";
+export { default as Footer } from "./Footer";
+export { default as BoardsFeed } from "./BoardsFeed";
+// export { default as TaskBoard } from "./TaskBoard";
+// export { default as EditableText } from "./EditableText";
+// export { default as DroppableZone } from "./DroppableZone";
